@@ -6,14 +6,14 @@ requires = [
 
 [project]
 name = "grudge"
-version = "2021.1"
+version = "2024.0"
 description = "Discretize discontinuous Galerkin operators quickly on heterogeneous hardware"
 readme = "README.rst"
 license = { text = "MIT" }
 authors = [
     { name = "Andreas Kloeckner", email = "inform@tiker.net" },
 ]
-requires-python = ">=3.8"
+requires-python = ">=3.10"
 classifiers = [
     "Development Status :: 3 - Alpha",
     "Intended Audience :: Developers",
@@ -125,4 +125,5 @@ known-local-folder = [
 lines-after-imports = 2
 
 [tool.mypy]
+python_version = "3.10"
 warn_unused_ignores = true
