@@ -42,8 +42,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 """
 
-
-from typing import Optional, Sequence
+from collections.abc import Sequence
 
 import numpy as np
 
@@ -71,7 +70,7 @@ from grudge.dof_desc import (
 
 def characteristic_lengthscales(
         actx: ArrayContext, dcoll: DiscretizationCollection,
-        dd: Optional[DOFDesc] = None) -> DOFArray:
+        dd: DOFDesc | None = None) -> DOFArray:
     r"""Computes the characteristic length scale :math:`h_{\text{loc}}` at
     each node. The characteristic length scale is mainly useful for estimating
     the stable time step size. E.g. for a hyperbolic system, an estimate of the
@@ -112,14 +111,15 @@ def characteristic_lengthscales(
                             cng * geo_facts
                             for cng, geo_facts in zip(
                                 dt_non_geometric_factors(dcoll, dd),
-                                actx.thaw(dt_geometric_factors(dcoll, dd)))))))
+                                actx.thaw(dt_geometric_factors(dcoll, dd)),
+                                strict=True)))))
 
     return actx.thaw(_compute_characteristic_lengthscales())
 
 
 @memoize_on_first_arg
 def dt_non_geometric_factors(
-        dcoll: DiscretizationCollection, dd: Optional[DOFDesc] = None
+        dcoll: DiscretizationCollection, dd: DOFDesc | None = None
         ) -> Sequence[float]:
     r"""Computes the non-geometric scale factors following [Hesthaven_2008]_,
     section 6.4, for each element group in the *dd* discretization:
@@ -142,7 +142,7 @@ def dt_non_geometric_factors(
     discr = dcoll.discr_from_dd(dd)
     min_delta_rs = []
     for grp in discr.groups:
-        nodes = np.asarray(list(zip(*grp.unit_nodes)))
+        nodes = np.asarray(list(zip(*grp.unit_nodes, strict=True)))
         nnodes = grp.nunit_dofs
 
         # NOTE: order 0 elements have 1 node located at the centroid of
@@ -169,8 +169,9 @@ def dt_non_geometric_factors(
 
 @memoize_on_first_arg
 def h_max_from_volume(
-        dcoll: DiscretizationCollection, dim=None,
-        dd: Optional[DOFDesc] = None) -> Scalar:
+        dcoll: DiscretizationCollection,
+        dim: int  | None = None,
+        dd: DOFDesc | None = None) -> Scalar:
     """Returns a (maximum) characteristic length based on the volume of the
     elements. This length may not be representative if the elements have very
     high aspect ratios.
@@ -201,8 +202,9 @@ def h_max_from_volume(
 
 @memoize_on_first_arg
 def h_min_from_volume(
-        dcoll: DiscretizationCollection, dim=None,
-        dd: Optional[DOFDesc] = None) -> Scalar:
+        dcoll: DiscretizationCollection,
+        dim: int | None = None,
+        dd: DOFDesc | None = None) -> Scalar:
     """Returns a (minimum) characteristic length based on the volume of the
     elements. This length may not be representative if the elements have very
     high aspect ratios.
@@ -232,7 +234,7 @@ def h_min_from_volume(
 
 
 def dt_geometric_factors(
-        dcoll: DiscretizationCollection, dd: Optional[DOFDesc] = None) -> DOFArray:
+        dcoll: DiscretizationCollection, dd: DOFDesc | None = None) -> DOFArray:
     r"""Computes a geometric scaling factor for each cell following
     [Hesthaven_2008]_, section 6.4, For simplicial elemenents, this factor is
     defined as the inradius (radius of an inscribed circle/sphere). For
@@ -392,6 +394,7 @@ def dt_geometric_factors(
                             actx.tag_axis(1, DiscretizationDOFAxisTag(), cv_i),
                             tagged=(FirstAxisIsElementsTag(),)) * r_fac
                         for cv_i, sae_i in zip(cell_vols, surface_areas)))))
+
 # }}}
 
 
