@@ -277,8 +277,10 @@ def _single_axis_derivative_kernel(
                                                get_diff_mat, vec_i, ijm_i,
                                                xyz_axis, metric_in_matvec)
             for out_grp, in_grp, vec_i, ijm_i in zip(
-                out_discr.groups, in_discr.groups, vec,
-                inv_jac_mat)))
+                out_discr.groups,
+                in_discr.groups,
+                vec,
+                inv_jac_mat, strict=True)))
 
 
 def _gradient_kernel(actx, out_discr, in_discr, get_diff_mat, inv_jac_mat, vec,
@@ -383,7 +385,7 @@ def _gradient_kernel(actx, out_discr, in_discr, get_diff_mat, inv_jac_mat, vec,
 
         for out_grp, in_grp, vec_i, ijm_i in zip(
             out_discr.groups, in_discr.groups, vec,
-            inv_jac_mat)
+                inv_jac_mat, strict=True)
     ]
 
     return make_obj_array([
@@ -506,7 +508,7 @@ def _divergence_kernel(actx, out_discr, in_discr, get_diff_mat, inv_jac_mat, vec
 
         for out_grp, in_grp, vec_i, ijm_i in zip(
             out_discr.groups, in_discr.groups, vec,
-            inv_jac_mat)
+                inv_jac_mat, strict=True)
     ]
 
     return DOFArray(actx, data=tuple(per_group_divs))
@@ -988,7 +990,8 @@ def _apply_mass_operator(
                 tagged=(FirstAxisIsElementsTag(),))
 
             for in_grp, out_grp, ae_i, vec_i in zip(
-                    in_discr.groups, out_discr.groups, area_elements, vec)
+                    in_discr.groups, out_discr.groups, area_elements, vec,
+                    strict=True)
         )
     )
 
@@ -1515,7 +1518,8 @@ def _apply_face_mass_operator(dcoll: DiscretizationCollection, dd_in, vec):
             for vgrp, afgrp, vec_i, surf_ae_i in zip(volm_discr.groups,
                                                      face_discr.groups,
                                                      vec,
-                                                     surf_area_elements)))
+                                                     surf_area_elements,
+                                                     strict=True)))
 
 
 def face_mass(dcoll: DiscretizationCollection, *args) -> ArrayOrContainer:
